Former-commit-id: 15520f7a32c79dd509f7c345b6c28e63f483086d

+++ Projects/UnitTests/Classes/Tests/SaveImageTest.cpp
@@ -36,7 +36,7 @@ using namespace DAVA;
 const float32 LOSSY_ALLOWED_DIFF = 2.f; //in percents
 const float32 LOSSLESS_ALLOWED_DIFF = 0.f; //in percents
 
-DAVA_TESTCLASS(SaveImageTest)
+DAVA_TESTCLASS (SaveImageTest)
 {
     Image* imageRGBA8888 = nullptr;
     Image* imageRGB888 = nullptr;
@@ -55,28 +55,28 @@ DAVA_TESTCLASS(SaveImageTest)
         SafeRelease(imageA8);
     }
 
-    DAVA_TEST(PngTest)
+    DAVA_TEST (PngTest)
     {
         SaveLoadCheck(imageRGBA8888, "testRGBA8888.png", LOSSLESS_ALLOWED_DIFF);
         //SaveLoadCheck(imageRGB888, "testRGB888.png", LOSSELESS_ALLOWED_DIFF); -- RGB888 is not supported by PNG
         SaveLoadCheck(imageA8, "testA8.png", LOSSLESS_ALLOWED_DIFF);
     }
 
-    DAVA_TEST(JpegTest)
+    DAVA_TEST (JpegTest)
     {
         //SaveLoadCheck(imageRGBA8888, "testRGBA8888.jpg", LOSSY_ALLOWED_DIFF); -- RGBA8888 is not supported for JPEG
         SaveLoadCheck(imageRGB888, "testRGB888.jpg", LOSSY_ALLOWED_DIFF);
         SaveLoadCheck(imageA8, "testA8.jpg", LOSSY_ALLOWED_DIFF);
     }
 
-    DAVA_TEST(TgaTest)
+    DAVA_TEST (TgaTest)
     {
         SaveLoadCheck(imageRGBA8888, "testRGBA8888.tga", LOSSLESS_ALLOWED_DIFF);
         SaveLoadCheck(imageRGB888, "testRGB888.tga", LOSSLESS_ALLOWED_DIFF);
         SaveLoadCheck(imageA8, "testA8.tga", LOSSLESS_ALLOWED_DIFF);
     }
 
-    DAVA_TEST(WebPTest)
+    DAVA_TEST (WebPTest)
     {
         SaveLoadCheck(imageRGB888, "testRGB888.webp", LOSSY_ALLOWED_DIFF);
         SaveLoadCheck(imageRGBA8888, "testRGBA8888.webp", LOSSY_ALLOWED_DIFF);
@@ -94,7 +94,7 @@ DAVA_TESTCLASS(SaveImageTest)
         TEST_VERIFY(imgSet[0]->dataSize == inImage->dataSize);
 
         const TextureUtils::CompareResult cmpRes = TextureUtils::CompareImages(inImage, imgSet[0], inImage->format);
-        float32 differencePersentage = ((float32)cmpRes.difference / ((float32)cmpRes.bytesCount * 256.f)) * 100.f;
+        float32 differencePersentage = (cmpRes.difference / (cmpRes.bytesCount * 256.f)) * 100.f;
         TEST_VERIFY(differencePersentage <= diffThreshold);
 
         for (auto img : imgSet)
